CowDetail/Cow Editing release
Stable version with decent looking CowDetail activity and a working edit
cow option.
