+++ VACAPP/.idea/libraries/play_services_measurement_8_1_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="play-services-measurement-8.1.0">
     <CLASSES>
-      <root url="file://$PROJECT_DIR$/app/build/intermediates/exploded-aar/com.google.android.gms/play-services-measurement/8.1.0/res" />
       <root url="jar://$PROJECT_DIR$/app/build/intermediates/exploded-aar/com.google.android.gms/play-services-measurement/8.1.0/jars/classes.jar!/" />
+      <root url="file://$PROJECT_DIR$/app/build/intermediates/exploded-aar/com.google.android.gms/play-services-measurement/8.1.0/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
