+++ VACAPP/app/build.gradle
@@ -21,7 +21,7 @@ android {
 
 dependencies {
     compile fileTree(include: ['*.jar'], dir: 'libs')
-    compile 'com.google.android.gms:play-services:8.1.0'
     compile project(':libraries:tess-two')
     compile project(':libraries:opencv')
+    compile 'com.google.android.gms:play-services:8.1.0'
 }
