+++ VACAPP/app/src/main/java/com/example/keinsfield/vacapp/Activities/CowDetailActivity.java
@@ -1,23 +1,20 @@
 package com.example.keinsfield.vacapp.Activities;
 
 import android.app.Activity;
-import android.app.AlertDialog;
-import android.app.Dialog;
 import android.app.ProgressDialog;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.pm.ActivityInfo;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
-import android.os.AsyncTask;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
 import android.widget.ArrayAdapter;
+import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ImageView;
 import android.widget.Spinner;
-import android.widget.TextView;
 
 import com.example.keinsfield.vacapp.Mundo.Cow;
 import com.example.keinsfield.vacapp.Mundo.CowKey;
@@ -25,31 +22,58 @@ import com.example.keinsfield.vacapp.Mundo.CowPersistenceManager;
 import com.example.keinsfield.vacapp.Mundo.Utilities;
 import com.example.keinsfield.vacapp.R;
 
-import java.io.BufferedReader;
 import java.io.File;
-import java.io.InputStreamReader;
-import java.net.URL;
 import java.util.ArrayList;
 import java.util.HashMap;
 
 public class CowDetailActivity extends Activity {
     private static final int PROGRESS_ID = 0;
-    private TextView nameText, textUltimoParto, textHato, textLoc, textPartos, textDiasLac, textLitros, textPrimerServicio;
+    private EditText nameText, textUltimoParto, textHato, textLoc, textPartos, textDiasLac, textLitros, textPrimerServicio;
+    private String lastName, lastUltimoP, lastHato, lastLoc, lastPartos, lastDiasLac, lastLitros, lastPrimerServ;
     private EditText cowNumberText;
+    private int currCowNumber;
+    private String currFinca;
     private Spinner spinner;
+    private ArrayAdapter<String> arrayAdapter;
+    private Button editButton, saveButton, cancelButton, findButton;
     private ImageView imageView;
     private ProgressDialog progressDialog;
-    private ArrayList<TextView> textViews;
+    private ArrayList<EditText> editTexts
+            ;
 
-    private void setUnknown() {
-        String s = "- -";
-        for (TextView view : textViews) view.setText(s);
+    private void setUnknown(boolean updateImage) {
+        String s = "";
+        for (EditText view : editTexts) view.setText(s);
+        if(!updateImage) return;
         BitmapFactory.Options opt = new BitmapFactory.Options();
         opt.inSampleSize = 4;
-        Bitmap bmp = BitmapFactory.decodeResource(getResources(),R.drawable.app_icon,opt);
+        Bitmap bmp = BitmapFactory.decodeResource(getResources(), R.drawable.app_icon,opt);
         imageView.setImageBitmap(bmp);
     }
 
+    private void disableAllTexts(){
+        for(EditText view: editTexts) {
+            disableView(view);
+        }
+    }
+
+    private void enableAllTexts(){
+        for(EditText view: editTexts){
+            enableView(view);
+        }
+    }
+
+    private void disableView(View view){
+        view.setEnabled(false);
+        //view.setFocusable(false);
+        //view.setClickable(false);
+    }
+
+    private void enableView(View view){
+        view.setEnabled(true);
+        //view.setFocusable(true);
+        //view.setClickable(true);
+    }
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -57,27 +81,41 @@ public class CowDetailActivity extends Activity {
         try {
             setRequestedOrientation(ActivityInfo.SCREEN_ORIENTATION_LANDSCAPE);
             imageView = (ImageView) findViewById(R.id.imageView);
-            nameText = (TextView) findViewById(R.id.nameText);
-            textUltimoParto = (TextView) findViewById(R.id.textLastParto);
-            textHato = (TextView) findViewById(R.id.textHato);
-            textLoc = (TextView) findViewById(R.id.textLoc);
-            textPartos = (TextView) findViewById(R.id.textPartos);
-            textDiasLac = (TextView) findViewById(R.id.textDiasLac);
-            textLitros = (TextView) findViewById(R.id.textLitros);
-            textPrimerServicio = (TextView) findViewById(R.id.textPrimerServ);
-            textViews = new ArrayList<>();
-            textViews.add(nameText);
-            textViews.add(textUltimoParto);
-            textViews.add(textHato);
-            textViews.add(textLoc);
-            textViews.add(textPartos);
-            textViews.add(textDiasLac);
-            textViews.add(textPrimerServicio);
-            setUnknown();
+            nameText = (EditText) findViewById(R.id.nameText);
+            textUltimoParto = (EditText) findViewById(R.id.textLastParto);
+            textHato = (EditText) findViewById(R.id.textHato);
+            textLoc = (EditText) findViewById(R.id.textLoc);
+            textPartos = (EditText) findViewById(R.id.textPartos);
+            textDiasLac = (EditText) findViewById(R.id.textDiasLac);
+            textLitros = (EditText) findViewById(R.id.textLitros);
+            textPrimerServicio = (EditText) findViewById(R.id.textPrimerServ);
+            editTexts = new ArrayList<>();
+            editTexts.add(nameText);
+            editTexts.add(textUltimoParto);
+            editTexts.add(textHato);
+            editTexts.add(textLoc);
+            editTexts.add(textPartos);
+            editTexts.add(textLitros);
+            editTexts.add(textDiasLac);
+            editTexts.add(textPrimerServicio);
             cowNumberText = (EditText) findViewById(R.id.cowNumberEditText);
             spinner = (Spinner) findViewById(R.id.farmSpinner);
-            ArrayAdapter<String> adapter = new ArrayAdapter<String>(this, android.R.layout.simple_spinner_item, Utilities.GetFarms(this));
-            spinner.setAdapter(adapter);
+            arrayAdapter = new ArrayAdapter(this, android.R.layout.simple_spinner_item, Utilities.GetFarms(this));
+            spinner.setAdapter(arrayAdapter);
+
+            editButton = (Button)findViewById(R.id.editButton);
+            saveButton = (Button)findViewById(R.id.saveButton);
+            cancelButton = (Button)findViewById(R.id.cancelButton);
+            findButton = (Button)findViewById(R.id.buttonFind);
+
+            saveButton.setVisibility(View.INVISIBLE);
+            disableView(saveButton);
+            cancelButton.setVisibility(View.INVISIBLE);
+            disableView(cancelButton);
+            editButton.setVisibility(View.INVISIBLE);
+            disableView(editButton);
+            setUnknown(true);
+            disableAllTexts();
 
         } catch (Exception e) {
             Log.d("SC", "FUCKED UP " + e.getMessage());
@@ -102,29 +140,53 @@ public class CowDetailActivity extends Activity {
         for (CowKey ckey : cows.keySet()) {
             Log.d("SC", "CKEY: " + ckey.nv + "," + ckey.farm + " " + ckey.equals(key));
         }
+        //If no cow is found, don't show the edit buttons.
         if (!cows.containsKey(key)) {
-            setUnknown();
+            setUnknown(true);
             Utilities.showDialog("Busqueda finalizada", "No se encontro ninguna vaca con numero " + nv + " en la finca " + finca + ".", this);
+            editButton.setVisibility(View.INVISIBLE);
+            disableView(editButton);
             return;
         }
+        disableAllTexts();
+        editButton.setVisibility(View.VISIBLE);
+        enableView(editButton);
+        currCowNumber = nv;
+        currFinca = finca;
         Cow cow = cows.get(key);
         updateCow(cow);
     }
 
     private void updateCow(Cow cow) {
-        if (cow.nombre.isEmpty()) nameText.setText("No name specificed.");
-        else nameText.setText(cow.nombre);
+        editButton.setVisibility(View.VISIBLE);
+        enableView(editButton);
+        setUnknown(true);
+        if (cow.nombre.isEmpty()) {
+            nameText.setText("No name specificed.");
+            lastName="";
+        }
+        else {
+            nameText.setText(cow.nombre);
+            lastName=cow.nombre;
+        }
         textUltimoParto.setText(cow.ultimo_parto);
-        String s = (cow.hato == -1) ? "- -" : cow.hato + "";
+        lastUltimoP = cow.ultimo_parto;
+        String s = (cow.hato == -1) ? "" : cow.hato + "";
         textHato.setText(s);
+        lastHato=s;
         textLoc.setText(cow.loc);
-        s = (cow.partos == -1) ? "- -" : cow.partos + "";
+        lastLoc=cow.loc;
+        s = (cow.partos == -1) ? "" : cow.partos + "";
         textPartos.setText(s);
-        s = (cow.dias_lac == -1) ? "- -" : cow.dias_lac + "";
+        lastPartos=s;
+        s = (cow.dias_lac == -1) ? "" : cow.dias_lac + "";
         textDiasLac.setText(s);
+        lastDiasLac=s;
         textPrimerServicio.setText(cow.primer_servicio);
-        s = (cow.lts_dia == -1) ? "- -" : cow.lts_dia + "";
+        lastPrimerServ=cow.primer_servicio;
+        s = (cow.lts_dia == -1) ? "" : cow.lts_dia + "";
         textLitros.setText(s);
+        lastLitros=s;
 
         //Look for image.
         File farm = new File(Utilities.GetStorageDirectory(this), cow.finca);
@@ -138,6 +200,130 @@ public class CowDetailActivity extends Activity {
                 }
             }
         }
+    }
+
+    private void refreshWithLastValues(){
+        cowNumberText.setText(currCowNumber+"");
+        for(int i = 0; i < arrayAdapter.getCount(); i ++){
+            if(arrayAdapter.getItem(i).equals(currFinca)){
+                spinner.setSelection(i);
+                break;
+            }
+        }
+        onFindClick(null);
+    }
+    private void afterModifySuccess(){
+        Utilities.showDialog("Vaca guardada exitosamente", "Vaca guardada exitosamente", this);
+        backToNormal();
+        refreshWithLastValues();
+    }
+
+    private void onCancelChanges(){
+        refreshWithLastValues();
+        backToNormal();
+    }
+
+    private void backToNormal(){
+        enableView(cowNumberText);
+        enableView(spinner);
+        disableAllTexts();
+        disableView(saveButton);
+        disableView(cancelButton);
+        saveButton.setVisibility(View.INVISIBLE);
+        cancelButton.setVisibility(View.INVISIBLE);
+        enableView(findButton);
+
+    }
+
+    public void onEditClick(View v){
+        enableAllTexts();
+        saveButton.setVisibility(View.VISIBLE);
+        enableView(saveButton);
+        cancelButton.setVisibility(View.VISIBLE);
+        disableView(editButton);
+        editButton.setVisibility(View.INVISIBLE);
+        enableView(cancelButton);
+        disableView(cowNumberText);
+        disableView(spinner);
+        disableView(findButton);
+    }
+
+    public void onSaveClick(View view){
+
+        try {
+            String s;
+            s = textHato.getText().toString();
+            int hato = s.isEmpty() ? -1 : Integer.parseInt(s);
+            s = textPartos.getText().toString();
+            int partos = s.isEmpty() ? -1 : Integer.parseInt(s);
+            s = textLitros.getText().toString();
+            int litros = s.isEmpty() ? -1 : Integer.parseInt(s);
+            s = textDiasLac.getText().toString();
+            int diasLac = s.isEmpty() ? -1 : Integer.parseInt(s);
+            String name = nameText.getText().toString();
+            String primer_serv = textPrimerServicio.getText().toString();
+            String loc = textLoc.getText().toString();
+            String ultimoP = textUltimoParto.getText().toString();
+            int number = currCowNumber;
+            String finca = currFinca;
+            final Cow newCow = new Cow(name, finca, number, ultimoP, hato, loc, partos, diasLac, litros, primer_serv);
+            DialogInterface.OnClickListener yesListener = new DialogInterface.OnClickListener() {
+                @Override
+                public void onClick(DialogInterface dialogInterface, int i) {
+                    try{
+                        CowPersistenceManager.modifyCow(newCow, CowDetailActivity.this);
+                        dialogInterface.dismiss();
+                        CowDetailActivity.this.afterModifySuccess();
+                    }
+                    catch(Exception e){
+                        Log.d("SC","Mistakes were made modifying cow.");
+                        e.printStackTrace();
+                    }
+                }
+            };
+            DialogInterface.OnClickListener noListener = new DialogInterface.OnClickListener() {
+                @Override
+                public void onClick(DialogInterface dialogInterface, int i) {
+                    try{
+                        return;
+                    }
+                    catch(Exception e){
+                    }
+                }
+            };
+            Utilities.showYesNoDialog("Guardar cambios","Esta seguro que desea guardar los cambios?", this, yesListener, "Guardar.", noListener, "Cancelar");
+
+        }
+        catch(Exception e){
+            Utilities.showDialog("Valores no validos.", "Existen valores no validos para algunos atributos.",this);
+            return;
+        }
+    }
+
+    public void onCancelClick(View view){
+        DialogInterface.OnClickListener yesListener = new DialogInterface.OnClickListener() {
+            @Override
+            public void onClick(DialogInterface dialogInterface, int i) {
+                try{
+                    onCancelChanges();
+                }
+                catch(Exception e){
+                    e.printStackTrace();
+                }
+            }
+        };
+        DialogInterface.OnClickListener noListener = new DialogInterface.OnClickListener() {
+            @Override
+            public void onClick(DialogInterface dialogInterface, int i) {
+                try{
+                    return;
+                }
+                catch(Exception e){
+                    e.printStackTrace();
+                }
+            }
+        };
+        Utilities.showYesNoDialog("Descartar cambios.","Seguro que desea descartar los cambios?",this,yesListener,"Si",noListener,"No");
     }
 
     public void BackToMain() {
