+++ VACAPP/app/src/main/res/values/colors.xml
@@ -24,4 +24,5 @@
     <color name="LightOrange">#ffc37e</color>
     <color name="LightBlue">#99E6FF</color>
     <color name="white">#ffffffff</color>
+    <color name="almost_black_gray">#03171B</color>
 </resources>
