+++ VACAPP/app/src/main/res/values/strings.xml
@@ -20,4 +20,5 @@
     <string name="fincasLocales">Fincas disponibles localmente:</string>
     <string name="title_activity_cow_detail">CowDetailActivity</string>
     <string name="title_activity_web_service">WebServiceActivity</string>
+    <string name="title_activity_maps">Map</string>
 </resources>
