+++ VACAPP/app/src/release/res/values/google_maps_api.xml
@@ -1,0 +1,21 @@
+<resources>
+    <!--
+    TODO: Before you run your application, you need a Google Maps API key.
+
+    To get one, follow this link, follow the directions and press "Create" at the end:
+
+    https://console.developers.google.com/flows/enableapi?apiid=maps_android_backend&keyType=CLIENT_SIDE_ANDROID&r=BB:18:92:2F:31:3A:52:CA:20:E9:9A:70:76:F9:19:66:2A:99:32:36%3Bcom.example.keinsfield.vacapp.Activities
+
+    You can also add your credentials to an existing key, using this line:
+    BB:18:92:2F:31:3A:52:CA:20:E9:9A:70:76:F9:19:66:2A:99:32:36;com.example.keinsfield.vacapp.Activities
+
+    Alternatively, follow the directions here:
+    https://developers.google.com/maps/documentation/android/start#get-key
+
+    Once you have your key (it starts with "AIza"), replace the "google_maps_key"
+    string in this file.
+    -->
+    <string name="google_maps_key" translatable="false" templateMergeStrategy="preserve">
+        YOUR_KEY_HERE
+    </string>
+</resources>
